feat: replace MultiBuf with reusable buffer
sodiumoxide 2.3 added the `pull_to_vec` and `push_to_vec` functions for
the `secretstream::Stream` type. This allows us to reuse allocated
memory instead of allocating on every crypto operation.

As a result of this change the existing MultiBuf solution can be removed
and replaced with a much more standard I/O using &[u8] buffers. This
also allowed the state machines to be cleaned up significantly.

Signed-off-by: Nick Stevens <nick@bitcurry.com>