ci: perform checks first, nightly last
Rearrange build order so that format and warning checks are performed
first and in their own stage. This way if format checks fail we don't
end up running the full build matrix.

Also move the nightly build to the end and add "fail_fast" so that the
overall build status can be determined before the nightly build (since
it can't affect the overall output anyway)

Signed-off-by: Nick Stevens <nick@bitcurry.com>