ut: fix transient failure of corrupt value test
The corrupt value test was directly writing a "0" to a byte to corrupt
the stream. However, it is totally possible that the value at that byte
is already zero and the stream is not actually corrupted, causing the
test to fail. This is fixed by incrementing the byte at that location by
one instead.

Signed-off-by: Nick Stevens <nick@bitcurry.com>